Update Route.php to allow attach with empty path
When using attach, it's not possible to add a route that uses the path_prefix provided in attach without a trailing slash. If the path itself is left blank, the path_prefix is ignored and the route used during matching is blank. With this modification, it will match on the route so for instance $router->attach('/blog', [ 'routes' => [ 'browse' => '' ] ]); will allow the route '/blog' to be matched. Without this, the route has to be '/blog/' with a trailing slash.