Adds functionality to identify requests with the X-Forwarded-Proto header set to https as secure
When a secure request is terminated at a load balancer and then forwarded to a server, it should
set the X-Forwarded-Proto header indicating that the original request used the https protocol.
Previously, the Secure Rule had two means of determining a request was secure $_SERVER['HTTPS'] == 'on'
and port = 443. This adds a check for $_SERVER['HTTP_X_FORWARDED_PROTO'] == 'https'.